refactor: #216 주문 상태 레이블/색상 상수 통합
@/constants/orderStatus (재export 배럴) 제거 후 모든 파일이
@/constants/status에서 직접 import하도록 통합
