--- src/constants/orderStatus.ts
@@ -1,1 +1,0 @@
-export { ORDER_STATUS_LABELS, ORDER_STATUS_COLORS } from './status';
